Do not ask to kill process we spawned

+++ terminal/terminal.go
@@ -132,13 +132,18 @@ func handleExit(client service.Client, t *Term) (error, int) {
 		}
 	}
 
-	answer, err := t.line.Prompt("Would you like to kill the process? [Y/n] ")
-	if err != nil {
-		return io.EOF, 2
-	}
-	answer = strings.ToLower(strings.TrimSpace(answer))
+	var kill bool
+	if client.AttachedToExistingProcess() {
+		answer, err := t.line.Prompt("Would you like to kill the process? [Y/n] ")
+		if err != nil {
+			return io.EOF, 2
+		}
+		answer = strings.ToLower(strings.TrimSpace(answer))
 
-	kill := (answer != "n" && answer != "no")
+		kill = (answer != "n" && answer != "no")
+	} else {
+		kill = true
+	}
 	err = client.Detach(kill)
 	if err != nil {
 		return err, 1
